# openapi/openapi.json
{
  "components": {
    "schemas": {
      "AdminOrderDto": {
        "properties": {
          "balance": {
            "minimum": 0,
            "type": "number"
          },
          "cancelledAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "orderNo": {
            "type": "string"
          },
          "pointsCostSnapshot": {
            "minimum": 1,
            "type": "number"
          },
          "productId": {
            "type": "string"
          },
          "productImageKeySnapshot": {
            "type": "string"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING_PICKUP",
              "COMPLETED",
              "CANCELLED"
            ],
            "type": "string"
          },
          "updatedBy": {
            "nullable": true,
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/OrderUserDto"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNo",
          "userId",
          "productId",
          "productNameSnapshot",
          "productImageKeySnapshot",
          "pointsCostSnapshot",
          "status",
          "createdAt",
          "completedAt",
          "cancelledAt",
          "updatedBy",
          "balance",
          "user"
        ],
        "type": "object"
      },
...
      "ApiErrorDto": {
        "properties": {
          "code": {
            "example": "VALIDATION_FAILED",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "example": "请求参数验证失败",
            "type": "string"
          },
          "requestId": {
            "example": "7da2aa93-ef82-45c8-9df0-84232e6a5b13",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId",
          "details"
        ],
        "type": "object"
      },
      "CreateOrderRequestDto": {
        "properties": {
          "productId": {
            "maxLength": 191,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "productId"
        ],
        "type": "object"
      },
...
      "OrderDto": {
        "properties": {
          "balance": {
            "minimum": 0,
            "type": "number"
          },
          "cancelledAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "orderNo": {
            "type": "string"
          },
          "pointsCostSnapshot": {
            "minimum": 1,
            "type": "number"
          },
          "productId": {
            "type": "string"
          },
          "productImageKeySnapshot": {
            "type": "string"
          },
          "productNameSnapshot": {
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING_PICKUP",
              "COMPLETED",
              "CANCELLED"
            ],
            "type": "string"
          },
          "updatedBy": {
            "nullable": true,
            "type": "string"
          },
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNo",
          "userId",
          "productId",
          "productNameSnapshot",
          "productImageKeySnapshot",
          "pointsCostSnapshot",
          "status",
          "createdAt",
          "completedAt",
          "cancelledAt",
          "updatedBy",
          "balance"
        ],
        "type": "object"
      },
      "OrderListResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/OrderDto"
            },
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
      "OrderUserDto": {
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username"
        ],
        "type": "object"
      },
      "PageMetaDto": {
        "properties": {
          "page": {
            "example": 1,
            "minimum": 1,
            "type": "number"
          },
          "pageSize": {
            "example": 20,
            "maximum": 100,
            "minimum": 1,
            "type": "number"
          },
          "total": {
            "example": 42,
            "minimum": 0,
            "type": "number"
          },
          "totalPages": {
            "example": 3,
            "minimum": 0,
            "type": "number"
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Android 与其他无 Cookie 客户端使用的访问令牌",
        "scheme": "bearer",
        "type": "http"
      },
      "cookieAuth": {
        "description": "Web 客户端访问令牌 Cookie",
        "in": "cookie",
        "name": "pq_access",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "英语答题、积分、商品兑换与订单管理开放接口",
    "title": "Point Quest API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/orders/{orderId}": {
      "get": {
        "operationId": "adminGetOrder",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOrderDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "获取订单管理详情",
        "tags": [
          "管理端-订单"
        ]
      }
    },
...
    "/api/v1/orders": {
      "get": {
        "operationId": "ordersList",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "分页查询我的订单",
        "tags": [
          "学生端-订单"
        ]
      },
      "post": {
        "operationId": "ordersCreate",
        "parameters": [
          {
            "description": "使用 Cookie 身份认证执行写操作时必填；Bearer 模式勿填",
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "同一用户内唯一，重试同一请求时必须复用",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "兑换商品并创建待领取订单",
        "tags": [
          "学生端-订单"
        ]
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "operationId": "ordersGet",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "maxLength": 191,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderDto"
                }
              }
            },
            "description": "成功"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求参数验证失败"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "身份认证失败"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "权限不足或 CSRF 校验失败"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "资源不存在"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "状态、幂等或并发冲突"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "请求体或上传文件过大"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorDto"
                }
              }
            },
            "description": "服务器内部错误"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "summary": "获取我的订单详情",
        "tags": [
          "学生端-订单"
        ]
      }
    },
...
  },
  "servers": [],
...
}